authorization

+++ controllers/feedController.js
@@ -3,6 +3,7 @@ const fs = require("fs");
 const path = require("path");
 
 const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -17,14 +18,12 @@ exports.getPosts = (req, res, next) => {
         .limit(perPage);
     })
     .then((posts) => {
-      res
-        .status(200)
-        .json({
-          posts: posts,
-          totalItems: totalItems,
-          perPage: perPage,
-          currentPage: currentPage,
-        });
+      res.status(200).json({
+        posts: posts,
+        totalItems: totalItems,
+        perPage: perPage,
+        currentPage: currentPage,
+      });
     })
     .catch((err) => {
       if (!err.statusCode) {
@@ -66,18 +65,31 @@ exports.createPost = (req, res, next) => {
     title: title,
     content: content,
     imageUrl: filePath,
-    creator: {
-      name: "Dennis",
-    },
+    creator: req.userId,
   });
+
+  let createdPost;
+  let creator;
 
   post
     .save()
     .then((result) => {
-      console.log(result);
+      createdPost = result;
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post has been created",
-        post: result,
+        post: createdPost,
+        creator: {
+          _id: creator._id,
+          name: creator.name,
+        },
       });
     })
     .catch((err) => {
@@ -134,6 +146,13 @@ exports.updatePost = (req, res, next) => {
         err.statusCode = 404;
         throw err;
       }
+
+      if (post.creator.toString() !== req.userId) {
+        const err = new Error("not authorized");
+        err.statusCode = 403;
+        throw err;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -169,12 +188,23 @@ exports.deletePost = (req, res, next) => {
       }
 
       //check logged in user
+      if (post.creator.toString() !== req.userId) {
+        const err = new Error("not authorized");
+        err.statusCode = 403;
+        throw err;
+      }
 
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
     })
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then((result) => {
       res.status(200).json({
         message: "Post has been deleted successfully!",
       });
